merge dev to main

+++ BE/src/main/java/com/example/onlinequiz/Repo/SliderRepository.java
@@ -13,5 +13,7 @@ public interface SliderRepository extends JpaRepository<Sliders, Long> {
 
     Sliders findBySliderID(int id);
 
-    void deleteBySliderID(int id);
+    void deleteBySliderID(long id);
+
+
 }
